// src_fatcAng/app/shared/data.dto.ts
export interface ICaseDTO {
    caseNumber: string;
    evidenceNumber: string;
    examinerName: string;
    description: string;
    id: number;
    ownerID: number;
}

export interface ICollectorTypes {
  name: string;
  author: string;
  version: string;
  metadata: string;
  //authURL: string;
  appID: string;
  authScopes: string;
}

export interface ICollectorAuthRequest {
  url: string;
  requestID: string;
}

export interface ICollectionSummaryItem {
  name: string;
  value: string;
}

export interface ICollectionSummary {
  id: number;
  caseId: number;
  collectorId: number;
  collectorSummary: { [itemName: string]: string; };
  retrievedAt: Date;
}

export interface ICollectorConfiguration {
  id: number;
  caseID: number;
  collectorTypeName: string;
  collectorAlias: string;
  hasAuthorization: boolean;
  authInfo: IAuthorization;
  desiredChunkSize: number;
  indexNumber: number;
  authRequestID: string;
  summary: ICollectionSummary;
}

export interface IAuthorization {
  username: string;
  password: string;
  authCode: string;
  authRequestID: string;
}
// ...
export interface IUserDTO {
  id: number;
  isAdmin: boolean;
  email: string;
  userName: string;
  firstName: string;
  lastName: string;
  passwordHash: string;
  authCode: string;
  tempPass: string;
  pwUserInput: string;
  configuration: string;
}

export interface IrunCollectorOption {
  parallel: boolean;
  series: boolean;
}

export const defaultCollector = {
  id: 0,
  caseID: 0,
  indexNumber: 0,
  collectorAlias: "",
  collectorTypeName: "",
  hasAuthorization: true,
  authCode: "",
  desiredChunkSize: 0,
  summary: {id: 0,
  caseId: 0,
  collectorId: 0,
  collectorSummary: null,
  retrievedAt: null}
};